vfo adjust
  * change increment to fractional values

+++ src/include/icom/IC705.h
@@ -181,7 +181,7 @@ public:
 
 	void setVfoAdj(double v);
 	double getVfoAdj();
-	void get_vfoadj_min_max_step(int &min, int &max, int &step) {
+	void get_vfoadj_min_max_step(double &min, double &max, double &step) {
 		min = 0; max = 100; step = 1; }
 
 	void set_band_selection(int v);
