+++ src/rigs/yaesu/FT991A.cxx
@@ -1451,7 +1451,7 @@ double RIG_FT991A::getVfoAdj()
 	return (double)(atoi(&replystr[p+5]));
 }
 
-void RIG_FT991A::get_vfoadj_min_max_step(int &min, int &max, int &step)
+void RIG_FT991A::get_vfoadj_min_max_step(double &min, double &max, double &step)
 {
 	min = -25;
 	max = 25;
