+++ src/rigs/yaesu/FTdx1200.cxx
@@ -353,7 +353,7 @@ double RIG_FTdx1200::getVfoAdj()
 	return (double)(atoi(&replystr[p+5]));
 }
 
-void RIG_FTdx1200::get_vfoadj_min_max_step(int &min, int &max, int &step)
+void RIG_FTdx1200::get_vfoadj_min_max_step(double &min, double &max, double &step)
 {
 	min = -25;
 	max = 25;
